pyproject.toml
[project]
name = "travel-rt"
version = "0.1.0"
description = ""
authors = [
    {name = "XetPy1030",email = "xetpy1030@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "psycopg-binary (>=3.2.6,<4.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-ckeditor-5 (>=0.2.17,<0.3.0)",
    "django-filter (>=25.1,<26.0)",
    "ruff (>=0.11.8,<0.12.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
